Split Type into a Typecast module
