+++ lib/artemis_ql/types.ex
@@ -1,11 +1,11 @@
 defmodule ArtemisQL.Types do
-  alias ArtemisQL.Types.ValueTransformError
   alias ArtemisQL.SearchMap
   alias ArtemisQL.Errors.KeyNotFound
   alias ArtemisQL.Errors.InvalidEnumValue
   alias ArtemisQL.Errors.UnsupportedSearchTermForField
 
   import ArtemisQL.Tokens
+  import ArtemisQL.Typecasts
 
   @doc """
   Attempts to filter the given key against the search_map, if the key is not in the map
@@ -571,185 +571,6 @@ defmodule ArtemisQL.Types do
 
   def normalize_value(value) do
     {:ok, value}
-  end
-
-  @spec cast_uuid(binary()) :: {:ok, String.t()} | :error
-  def cast_uuid(str) do
-    Ecto.UUID.cast(str)
-  end
-
-  @spec cast_ulid(binary()) :: {:ok, String.t()} | :error
-  def cast_ulid(str) do
-    Ecto.ULID.cast(str)
-  end
-
-  @spec cast_boolean(String.t()) :: {:ok, boolean()}
-  def cast_boolean(value) when is_binary(value) do
-    case String.downcase(value) do
-      v when v in ~w[yes y 1 true t on] ->
-        {:ok, true}
-
-      v when v in ~w[no n 0 false f off] ->
-        {:ok, false}
-
-      _ ->
-        :error
-    end
-  end
-
-  @spec cast_integer(String.t()) :: {:ok, integer()} | :error
-  def cast_integer(value) when is_integer(value) do
-    {:ok, value}
-  end
-
-  def cast_integer(value) when is_binary(value) do
-    {sign, value} =
-      case value do
-        <<"-", rest::binary>> -> {-1, rest}
-        <<"+", rest::binary>> -> {1, rest}
-        _ -> {1, value}
-      end
-
-    with {:ok, value} <- parse_integer_with_optional_base(value) do
-      {:ok, value * sign}
-    else
-      :error ->
-        :error
-    end
-  end
-
-  def cast_float(value) when is_float(value), do: {:ok, value}
-  def cast_float(value) when is_integer(value), do: {:ok, value * 1.0}
-
-  def cast_float(value) when is_binary(value) do
-    with {:ok, value} <- normalize_decimal_like_numeric_literal(value) do
-      Ecto.Type.cast(:float, value)
-    else
-      :error ->
-        :error
-    end
-  end
-
-  def cast_decimal(%Decimal{} = value), do: {:ok, value}
-
-  def cast_decimal(value) when is_integer(value) or is_float(value) do
-    Ecto.Type.cast(:decimal, value)
-  end
-
-  def cast_decimal(value) when is_binary(value) do
-    with {:ok, value} <- normalize_decimal_like_numeric_literal(value) do
-      Ecto.Type.cast(:decimal, value)
-    else
-      :error ->
-        :error
-    end
-  end
-
-  defp parse_integer_with_optional_base(<<"0b", value::binary>>), do: parse_integer_by_base(value, 2)
-  defp parse_integer_with_optional_base(<<"0B", value::binary>>), do: parse_integer_by_base(value, 2)
-  defp parse_integer_with_optional_base(<<"0o", value::binary>>), do: parse_integer_by_base(value, 8)
-  defp parse_integer_with_optional_base(<<"0O", value::binary>>), do: parse_integer_by_base(value, 8)
-  defp parse_integer_with_optional_base(<<"0x", value::binary>>), do: parse_integer_by_base(value, 16)
-  defp parse_integer_with_optional_base(<<"0X", value::binary>>), do: parse_integer_by_base(value, 16)
-  defp parse_integer_with_optional_base(value), do: parse_integer_by_base(value, 10)
-
-  defp parse_integer_by_base("", _base), do: :error
-
-  defp parse_integer_by_base(value, base) do
-    with {:ok, value} <- normalize_integer_literal(value, base) do
-      case Integer.parse(value, base) do
-        {value, ""} ->
-          {:ok, value}
-
-        _ ->
-          :error
-      end
-    else
-      :error ->
-        :error
-    end
-  end
-
-  defp normalize_integer_literal(value, base) when is_binary(value) do
-    validate_and_normalize_numeric_literal(value, fn c ->
-      c in integer_digit_chars_for_base(base)
-    end)
-  end
-
-  defp normalize_decimal_like_numeric_literal(value) when is_binary(value) do
-    validate_and_normalize_numeric_literal(value, fn c ->
-      c >= ?0 and c <= ?9
-    end)
-  end
-
-  defp validate_and_normalize_numeric_literal(value, digit_fun) when is_binary(value) do
-    chars = :binary.bin_to_list(value)
-
-    case chars do
-      [] ->
-        :error
-
-      _ ->
-        valid? =
-          Enum.with_index(chars)
-          |> Enum.all?(fn
-            {?_, index} ->
-              index > 0 and
-                index < length(chars) - 1 and
-                digit_fun.(Enum.at(chars, index - 1)) and
-                digit_fun.(Enum.at(chars, index + 1))
-
-            {_c, _index} ->
-              true
-          end)
-
-        case valid? do
-          true ->
-            {:ok, String.replace(value, "_", "")}
-
-          false ->
-            :error
-        end
-    end
-  end
-
-  defp integer_digit_chars_for_base(2), do: ~c"01"
-  defp integer_digit_chars_for_base(8), do: ~c"01234567"
-  defp integer_digit_chars_for_base(10), do: ~c"0123456789"
-  defp integer_digit_chars_for_base(16), do: ~c"0123456789abcdefABCDEF"
-  defp integer_digit_chars_for_base(_base), do: ~c"0123456789"
-
-  @spec cast_atom(String.t()) :: {:ok, atom()}
-  def cast_atom(str) do
-    {:ok, String.to_existing_atom(str)}
-  end
-
-  @spec cast_date(String.t()) :: {:ok, Date.t()}
-  def cast_date(value) do
-    {:ok, ArtemisQL.Types.DateAndTime.parse_date(value)}
-  rescue _ex in ValueTransformError ->
-    :error
-  end
-
-  @spec cast_time(String.t()) :: {:ok, Time.t()}
-  def cast_time(value) do
-    {:ok, ArtemisQL.Types.DateAndTime.parse_time(value)}
-  rescue _ex in ValueTransformError ->
-    :error
-  end
-
-  @spec cast_datetime(String.t()) :: {:ok, Time.t()}
-  def cast_datetime(value) do
-    {:ok, ArtemisQL.Types.DateAndTime.parse_datetime(value)}
-  rescue _ex in ValueTransformError ->
-    :error
-  end
-
-  @spec cast_naive_datetime(String.t()) :: {:ok, Time.t()}
-  def cast_naive_datetime(value) do
-    {:ok, ArtemisQL.Types.DateAndTime.parse_naive_datetime(value)}
-  rescue _ex in ValueTransformError ->
-    :error
   end
 
   #
